refactor: update module imports to align with new directory structure
