+++ apps/electron/src/main/modules/agent/deployed/deployed-agent.ts
@@ -12,7 +12,7 @@ import {
   connectToMCPServer,
   fetchServerTools,
   substituteArgsParameters,
-} from "@/main/modules/mcp-core/apps/mcp-apps-service";
+} from "@/main/modules/mcp-apps-manager/mcp-apps-service";
 
 /**
  * サーバー情報を表す型
