+++ apps/electron/src/main/modules/mcp-apps-manager/app-paths.ts
