+++ apps/electron/src/main/modules/mcp-apps-manager/mcp-client.ts
