+++ apps/electron/src/main/modules/mcp-apps-manager/mcp-config-importer.ts
@@ -1,5 +1,5 @@
 import { promises as fsPromises } from "fs";
-import { getServerService } from "@/main/modules/mcp-core/mcp-manager/server-service";
+import { getServerService } from "@/main/modules/mcp-server-manager/server-service";
 import { MCPServerConfig, ClientType, ClientConfig } from "@mcp_router/shared";
 import {
   claudeConfig,
