+++ apps/electron/src/main/modules/mcp-apps-manager/token-manager.ts
