+++ apps/electron/src/main/modules/mcp-server-manager/aggregator-server.ts
