+++ apps/electron/src/main/modules/mcp-server-manager/dxt-processor/dxt_schema.md
