+++ apps/electron/src/main/modules/mcp-server-manager/index.ts
