+++ apps/electron/src/main/modules/mcp-server-manager/request-handler-base.ts
@@ -37,10 +37,10 @@ export abstract class RequestHandlerBase {
     try {
       // WorkflowServiceとWorkflowExecutorをインポート
       const { getWorkflowService } = await import(
-        "../../workflow/workflow.service"
+        "../workflow/workflow.service"
       );
       const { WorkflowExecutor } = await import(
-        "../../workflow/workflow-executor"
+        "../workflow/workflow-executor"
       );
       const workflowService = getWorkflowService();
