+++ apps/electron/src/main/modules/mcp-server-manager/request-handlers.ts
