+++ apps/electron/src/main/ui/tray.ts
@@ -1,5 +1,5 @@
 import { app, Menu, Tray, nativeImage } from "electron";
-import { MCPServerManager } from "@/main/modules/mcp-core/mcp-manager";
+import { MCPServerManager } from "@/main/modules/mcp-server-manager";
 import { mainWindow } from "../../main";
 
 // Global tray instance
